Move output tests to separate module
